Fix: Display error when the email is already in use.
Edit the .subscribe of the signup front end, so that it displays an alert with the error. The console still logs the error, as apparently there is no way around it. People are welcome to check if there is a way, but for now, there isn't a problem with the functionality of the code. Darwish was consulted for this decision.