fix: Show user info in header.
In the latest commits after changing the routes. the user info didn't
show up as a result of  following one time evaluation technique
OnInit for the menu items of the login page. The chagnes introduced
added the capability for continuous checks over the user stored in the
user service, using *ngIf statements on the nb-user component inside the
header component.